Adding specs
These are just some basic feature tests mostly focused on visiting pages and checking content. There are a couple of more advanced tests relevant to this project to test for the presence of authenticity token fields and for proper DELETE requests.

One thing to note, is that checking for `authenticity_token` fails if `form_with` is used. 

That sounds a little sketchy, but the point of this specific project is for someone to hand-write the `authenticity_token` code. So it's good to check for that, and actually _good_ that this doesn't pass with `form_with`.